+++ ChangeLog.md
@@ -20,7 +20,7 @@
 - camel-case of some property names
 - Text instances for "Base64" moved to `Data.TypedEncoding.Instances.Enc.Warn.Base64`
 - Removed instanced for `"r-()"` encoding
-
+- Fuctions from `Data.TypedEncoding.Instances.Support.Common` moved to `Data.TypedEncoding.Instances.Support.Decode`
 - More general instances for some encodings in `Data.TypedEncoding.Instances.Restriction.Misc`
 
 ## 0.4.2